#64 処理の分離を行った
