+++ src/features/chat/hooks/useSendChat.ts
@@ -1,5 +1,3 @@
-import { axiosInstance } from '@/lib/axios';
-import { useMutation } from '@tanstack/react-query';
 import { useCallback, useState } from 'react';
 import { postChatSSE } from '../api/postChatSSE';
 
@@ -20,42 +18,16 @@ export const useSendChat = () => {
   const sendChat = useCallback(
     async ({ message, threadId }: SendChatArgument) => {
       setLatestQuestion(message);
-
-      const optionalQuery = threadId ? `&thread_id=${threadId}` : '';
       let newThreadId: string | undefined = undefined;
 
       try {
-        await axiosInstance.get(`/chat/sse?text=${message}${optionalQuery}`, {
-          headers: {
-            'Content-Type': 'application/json',
-            Accept: 'text/event-stream',
-          },
-          responseType: 'stream',
-          onDownloadProgress: (progressEvent) => {
-            const chunk = progressEvent.event.target.response;
-            if (!chunk) return;
-
-            const lines = chunk
-              .split('data: ')
-              .map((line: string) => line.trim())
-              .filter((s: string) => s);
-
-            for (const json of lines) {
-              try {
-                if (json === '[DONE]') {
-                  return;
-                }
-                const data = JSON.parse(json);
-                const text = data.content;
-
-                setLatestAnswer(text);
-
-                if (!newThreadId && data.thread_id) {
-                  newThreadId = data.thread_id;
-                }
-              } catch (error) {
-                console.error('Error parsing JSON:', error);
-              }
+        await postChatSSE({
+          message,
+          threadId,
+          onMessage: (text, threadId) => {
+            setLatestAnswer(text);
+            if (!newThreadId && threadId) {
+              newThreadId = threadId;
             }
           },
         });
